Close final review backlog: rational fps, shared ffmpeg helper, explicit imports
CR-L2: probe keeps the fractional frame rate (e.g. 23.976) and Plan.fps is float,
so long clips no longer drift from an int-rounded 24 fps.

CR-L4: new app/pipeline/_ff.run captures stderr and raises FFmpegError with the
stderr tail; demux/audio_mix/branding/video_ops/metadata adopt it instead of bare
check=True that hid ffmpeg diagnostics.

CR-L9 / TD-1: replace dual-context try/except imports (which swallowed real
ImportErrors) with explicit __package__ detection (api, metadata) and relative
sibling imports (tighten, segment). Verified in both package and flat/server
import contexts.

tests/test_batch.py +2 (ffmpeg-error surfacing, no-import-swallow). Full suite
12/12 green. Whole CODE-REVIEW backlog now Fixed or verified-correct: 0 open.
DECISIONS.md I-7 + TD-1 Closed.

+++ reelcut/app/pipeline/render.py
@@ -74,7 +74,7 @@ class Plan:
     boundaries: List[Boundary]
     width: int = 1280
     height: int = 720
-    fps: int = 30
+    fps: float = 30.0          # may be fractional (e.g. 23.976) — CR-L2
     timing_map: List[dict] = field(default_factory=list)
 
 
@@ -126,7 +126,7 @@ def build_plan(project: dict) -> Plan:
             auto_gap=not was_adjacent))
     return Plan(project["source"], clips, boundaries,
                 int(project.get("width", 1280)), int(project.get("height", 720)),
-                int(project.get("fps", 30)))
+                float(project.get("fps", 30)))
 
 
 def timing_map(plan: Plan) -> List[dict]:
